generate/requirements.txt
accelerate==0.18.0
huggingface_hub==0.14.1
jieba==0.42.1
nltk==3.8.1
numpy==1.24.3
rouge_chinese==1.0.3 
sentencepiece==0.1.99
tqdm==4.65.0
transformers==4.28.1
pytorch==2.0.1 
torchvision==0.15.2 
cudatoolkit==11.8
torchaudio==2.0.2 
tqdm==4.65.0
fastapi==0.110.2 